# shared/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LibiaoLink API（市场项目管理系统）",
    "version": "0.1.0",
    "description": "契约唯一真相 = shared/src 下的 Zod schema（技术设计v0.2 §7）；本文件与 generated/api-types.d.ts 由工具生成，禁止手改（CONTRIBUTING §14）。"
  },
  "servers": [
    {
      "url": "/",
      "description": "同源部署（Nginx 反代 /api/v1）"
    }
  ],
...
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string",
            "example": "参数校验失败"
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "default": []
          },
          "traceId": {
            "type": "string",
            "example": "4f1c2f2e-6f8a-4b1e-9a1f-2f6d6f2c9d10"
          }
        },
        "required": [
          "code",
          "message",
          "traceId"
        ],
        "description": "统一错误信封（技术设计v0.2 §7.2）"
      },
      "Blueprint": {
        "type": "object",
        "properties": {
          "schemaVersion": {
            "type": "number",
            "enum": [
              1
            ]
          },
          "blueprintVersion": {
            "type": "integer",
            "exclusiveMinimum": 0
          },
          "name": {
            "type": "string"
          },
          "projectType": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "所属项目类型（ADR-019：按项目类型各一份，default 为兜底模板）；缺省 = 通用"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/DateTime"
          },
          "stages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlueprintStage"
            },
            "minItems": 1
          }
        },
        "required": [
          "schemaVersion",
          "blueprintVersion",
          "name",
          "updatedAt",
          "stages"
        ],
        "description": "自建蓝图 JSON（v0.2 §3.2 + ADR-019）；导入校验 = schema + 引用 + 幂等"
      },
      "BlueprintConstraint": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/BlueprintRequiredDocConstraint"
          },
          {
            "$ref": "#/components/schemas/BlueprintReservedConstraint"
          }
        ]
      },
      "BlueprintImportBody": {
        "type": "object",
        "properties": {
          "blueprint": {
            "$ref": "#/components/schemas/Blueprint"
          }
        },
        "required": [
          "blueprint"
        ],
        "description": "导入自建蓝图 JSON（保存为草稿；重复导入幂等）"
      },
      "BlueprintNode": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "example": "design.mech",
            "description": "节点稳定键；一经发布不可改名（改名 = 新增 + 废弃）"
          },
          "name": {
            "type": "string",
            "example": "机械设计图纸"
          },
          "seq": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "description": "排序（建议 10/20/30 步长，便于插入）"
          },
          "constraints": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlueprintConstraint"
            },
            "default": []
          }
        },
        "required": [
          "key",
          "name",
          "seq"
        ],
        "description": "每个节点至少一类约束（校验规则见 v0.2 §3.4）"
      },
      "BlueprintRequiredDocConstraint": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "required_doc"
            ]
          },
          "docType": {
            "$ref": "#/components/schemas/DocType"
          },
          "minCount": {
            "type": "integer",
            "minimum": 1,
            "default": 1
          }
        },
        "required": [
          "type",
          "docType"
        ],
        "description": "必交成果物（一期门禁实现的唯一约束类型）"
      },
      "BlueprintReservedConstraint": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "field",
              "dependency",
              "deadline"
            ]
          },
          "config": {
            "type": "object",
            "additionalProperties": {},
            "default": {}
          }
        },
        "required": [
          "type"
        ],
        "description": "预留约束类型（结构就位、规则后置）"
      },
      "BlueprintSaveBody": {
        "type": "object",
        "properties": {
          "blueprint": {
            "$ref": "#/components/schemas/Blueprint"
          }
        },
        "required": [
          "blueprint"
        ],
        "description": "保存蓝图草稿（必须通过 schema + 引用校验）"
      },
      "BlueprintStage": {
        "type": "object",
        "properties": {
          "key": {
            "$ref": "#/components/schemas/StageKey"
          },
          "name": {
            "type": "string"
          },
          "seq": {
            "type": "integer",
            "exclusiveMinimum": 0
          },
          "nodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlueprintNode"
            },
            "minItems": 1
          }
        },
        "required": [
          "key",
          "name",
          "seq",
          "nodes"
        ]
      },
      "BlueprintStatus": {
        "type": "string",
        "enum": [
          "draft",
          "published"
        ],
        "description": "蓝图状态：发布产生新版本，不自动影响已生成项目（快照）"
      },
      "BlueprintView": {
        "type": "object",
        "properties": {
          "blueprint": {
            "$ref": "#/components/schemas/Blueprint"
          },
          "status": {
            "$ref": "#/components/schemas/BlueprintStatus"
          },
          "projectType": {
            "type": "string",
            "description": "所属项目类型（ADR-019）"
          },
          "isDefault": {
            "type": "boolean",
            "description": "是否兜底默认模板（project_type = default）"
          },
          "publishedVersion": {
            "type": "integer",
            "minimum": 0,
            "description": "已发布版本号（0 = 尚未发布）"
          },
          "publishedAt": {
            "allOf": [
              {
                "$ref": "#/components/schemas/DateTime"
              },
              {
                "type": [
                  "string",
                  "null"
                ]
              }
            ]
          },
          "publishedBy": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Uuid"
              },
              {
                "type": [
                  "string",
                  "null"
                ]
              }
            ]
          }
        },
        "required": [
          "blueprint",
          "status",
          "projectType",
          "isDefault",
          "publishedVersion",
          "publishedAt",
          "publishedBy"
        ]
      },
...
      "DateTime": {
        "type": "string",
        "format": "date-time",
        "description": "ISO8601 时间戳（UTC 存储，前端按 Asia/Shanghai 展示）"
      },
...
      "DocType": {
        "type": [
          "string",
          "null"
        ],
        "enum": [
          "CAD图纸",
          "技术协议",
          "合同",
          "评审单",
          "设备清单",
          "物料总清单",
          "发货装箱单",
          "到货单",
          "安装完成证明",
          "验收单",
          null
        ],
        "description": "十类成果文件字典；门禁 required_doc 只能引用此字典（v0.2 §2.5）"
      },
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "VALIDATION_FAILED",
          "AUTH_REQUIRED",
          "AUTH_CALLBACK_FAILED",
          "FORBIDDEN",
          "NOT_FOUND",
          "VERSION_CONFLICT",
          "PROJECT_CODE_EXISTS",
          "PROJECT_ARCHIVED",
          "STAGE_GATE_NOT_PASSED",
          "BLUEPRINT_NOT_PUBLISHED",
          "NODE_REQUIRED_DOC_MISSING",
          "NODE_HAS_FILES",
          "STAGE_STATE_INVALID",
          "NODE_ALREADY_DONE",
          "NODE_ALREADY_EXISTS",
          "NODE_DELETED",
          "TASK_ALREADY_EXISTS",
          "BLUEPRINT_SCHEMA_INVALID",
          "BLUEPRINT_REF_UNKNOWN",
          "FILE_STATE_INVALID",
          "UPLOAD_INCOMPLETE",
          "UPLOAD_SESSION_EXPIRED",
          "FILE_HASH_MISMATCH",
          "IDEMPOTENT_REPLAY",
          "PREVIEW_NOT_READY",
          "PREVIEW_FAILED",
          "INTERNAL"
        ],
        "description": "统一错误码（技术设计v0.2 §7.2）"
      },
      "ErrorDetail": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "too_small"
          },
          "message": {
            "type": "string",
            "example": "limit 必须是 1~200 的整数"
          },
          "path": {
            "type": "string",
            "example": "query.limit"
          },
          "meta": {
            "type": "object",
            "additionalProperties": {}
          }
        },
        "required": [
          "code",
          "message"
        ],
        "description": "字段级错误明细（校验失败、门禁缺件等）"
      },
...
      "IdempotencyKey": {
        "type": "string",
        "minLength": 8,
        "maxLength": 128,
        "description": "写操作幂等键（Idempotency-Key 请求头）；重复提交返回首次结果"
      },
...
      "StageKey": {
        "type": "string",
        "enum": [
          "presale",
          "design",
          "purchase",
          "assembly",
          "install",
          "deploy",
          "trial",
          "production",
          "acceptance"
        ],
        "description": "九阶段字典（v0.2 §2.5：售前规划 / 设计开发 / 加工采购 / 组装发货 / 硬件实施 / 软件部署 / 试运行 / 生产阶段 / 验收）"
      },
...
      "Uuid": {
        "type": "string",
        "format": "uuid",
        "description": "UUID（主键与关联 ID）"
      },
...
    },
...
  },
  "paths": {
...
    "/api/v1/blueprint": {
      "get": {
        "tags": [
          "flow"
        ],
        "summary": "当前蓝图（含版本与发布状态；该项目类型尚未建档时 404，default 仅作导入兜底）",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "description": "项目类型（ADR-019）；缺省 = default 兜底模板"
            },
            "required": false,
            "description": "项目类型（ADR-019）；缺省 = default 兜底模板",
            "name": "projectType",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "蓝图视图",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlueprintView"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在或不可见（NOT_FOUND，统一 404 语义）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "flow"
        ],
        "summary": "保存蓝图草稿（必须通过 schema + 引用校验）",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "description": "项目类型（ADR-019）；缺省 = default 兜底模板"
            },
            "required": false,
            "description": "项目类型（ADR-019）；缺省 = default 兜底模板",
            "name": "projectType",
            "in": "query"
          },
          {
            "schema": {
              "$ref": "#/components/schemas/IdempotencyKey"
            },
            "required": false,
            "description": "写操作幂等键（Idempotency-Key 请求头）；重复提交返回首次结果",
            "name": "Idempotency-Key",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlueprintSaveBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "保存后的蓝图视图",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlueprintView"
                }
              }
            }
          },
          "400": {
            "description": "契约校验失败（VALIDATION_FAILED）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "无权限（FORBIDDEN）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "业务校验未通过（门禁 / 蓝图校验，含明细）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/blueprint/publish": {
      "post": {
        "tags": [
          "flow"
        ],
        "summary": "发布蓝图（递增 blueprintVersion；不影响已生成项目）",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "description": "项目类型（ADR-019）；缺省 = default 兜底模板"
            },
            "required": false,
            "description": "项目类型（ADR-019）；缺省 = default 兜底模板",
            "name": "projectType",
            "in": "query"
          },
          {
            "schema": {
              "$ref": "#/components/schemas/IdempotencyKey"
            },
            "required": false,
            "description": "写操作幂等键（Idempotency-Key 请求头）；重复提交返回首次结果",
            "name": "Idempotency-Key",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlueprintSaveBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "发布后的蓝图视图",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlueprintView"
                }
              }
            }
          },
          "403": {
            "description": "无权限（FORBIDDEN）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在或不可见（NOT_FOUND，统一 404 语义）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "业务校验未通过（门禁 / 蓝图校验，含明细）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/blueprint/export": {
      "get": {
        "tags": [
          "flow"
        ],
        "summary": "导出蓝图 JSON（自建格式，round-trip 无损）",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "description": "项目类型（ADR-019）；缺省 = default 兜底模板"
            },
            "required": false,
            "description": "项目类型（ADR-019）；缺省 = default 兜底模板",
            "name": "projectType",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "蓝图 JSON",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Blueprint"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在或不可见（NOT_FOUND，统一 404 语义）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/blueprint/import": {
      "post": {
        "tags": [
          "flow"
        ],
        "summary": "导入蓝图 JSON（保存为草稿；重复导入幂等）",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "description": "项目类型（ADR-019）；缺省 = default 兜底模板"
            },
            "required": false,
            "description": "项目类型（ADR-019）；缺省 = default 兜底模板",
            "name": "projectType",
            "in": "query"
          },
          {
            "schema": {
              "$ref": "#/components/schemas/IdempotencyKey"
            },
            "required": false,
            "description": "写操作幂等键（Idempotency-Key 请求头）；重复提交返回首次结果",
            "name": "Idempotency-Key",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlueprintImportBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "导入后的蓝图视图",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlueprintView"
                }
              }
            }
          },
          "403": {
            "description": "无权限（FORBIDDEN）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "业务校验未通过（门禁 / 蓝图校验，含明细）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}